ml-server/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
numpy
pandas
scikit-learn
xgboost
redis
python-multipart
httpx
python-dotenv
kafka-python
langchain
langchain-core
langchain-openai
langgraph
openai
joblib
mlflow>=2.9.0
optuna>=3.4.0
shap>=0.43.0
prometheus-fastapi-instrumentator>=6.1.0
SpeechRecognition==3.10.0
pydub==0.25.1
openai-whisper==20231117